+++ metrics-core/src/main/java/com/yammer/metrics/core/Timer.java
@@ -3,10 +3,6 @@ package com.yammer.metrics.core;
 import com.yammer.metrics.core.Histogram.SampleType;
 import com.yammer.metrics.stats.Snapshot;
 
-import java.io.File;
-import java.io.IOException;
-import java.util.ArrayList;
-import java.util.List;
 import java.util.concurrent.Callable;
 import java.util.concurrent.ScheduledExecutorService;
 import java.util.concurrent.TimeUnit;
@@ -185,29 +181,6 @@ public class Timer implements Metered, Stoppable, Sampling, Summarizable {
     @Override
     public String eventType() {
         return meter.eventType();
-    }
-
-    /**
-     * Returns a list of all recorded durations in the timer's sample.
-     *
-     * @return a list of all recorded durations in the timer's sample
-     */
-    public List<Double> values() {
-        final List<Double> values = new ArrayList<Double>();
-        for (Long value : histogram.values()) {
-            values.add(convertFromNS(value));
-        }
-        return values;
-    }
-
-    /**
-     * Writes the values of the timer's sample to the given file.
-     *
-     * @param output the file to which the values will be written
-     * @throws java.io.IOException if there is an error writing the values
-     */
-    public void dump(File output) throws IOException {
-        histogram.dump(output);
     }
 
     private void update(long duration) {