+++ metrics-core/src/main/java/com/yammer/metrics/stats/Sample.java
@@ -1,9 +1,5 @@
 package com.yammer.metrics.stats;
 
-import java.io.File;
-import java.io.IOException;
-import java.util.List;
-
 /**
  * A statistically representative sample of a data stream.
  */
@@ -33,19 +29,4 @@ public interface Sample {
      * @return a snapshot of the sample's values
      */
     Snapshot getSnapshot();
-
-    /**
-     * Returns a copy of the sample's values.
-     *
-     * @return a copy of the sample's values
-     */
-    List<Long> values();
-
-    /**
-     * Writes the values of the sample to the given file.
-     *
-     * @param output the file to which the values will be written
-     * @throws IOException if there is an error writing the values
-     */
-    void dump(File output) throws IOException;
 }